Rewrote most mem* and str* function in Assembly, except for strcmp for optimisations

+++ kernel/src/lib/stdint.h
@@ -7,8 +7,8 @@ typedef unsigned short      uint16_t;
 typedef          short       int16_t;
 typedef unsigned int        uint32_t;
 typedef          int         int32_t;
-typedef unsigned long long  uint64_t;
-typedef          long long   int64_t;
+typedef unsigned long       uint64_t;
+typedef          long        int64_t;
 
 typedef uint64_t size_t;
